requirements.txt
Django==2.2.10
djangorestframework==3.11.1
pytz==2020.1
sqlparse==0.3.1
